miatoll: Disable VSync for CPU rendered apps
- Choreographer VSync has some performance impact for CPU rendered apps
as the app has to wait for VSync to start a new frame draw. When the
frame draw time exceeds one VSync, this has a direct impact on the FPS.
"debug.cpurend.vsync" property controls the VSync for CPU rendered apps.

Disable VSync for CPU rendered app, by setting this property to false.

Change-Id: I93cfa7c588f1ab80803940c8059073075fd49506